Balance Update + Raycasting Bug Fixes
Fixed bug with main menu not working after going back from the pause menu (issue was that the time was not continued).
Now updating the name of the currently active character.
Balance is being updated, decreased if character is carrying something in spaced and moving.
Raycasting issue fixed: now player can regrab items inside the box, and pizza 02 is being set properly